+++ src/main/kotlin/com/github/nadiatarashkevich/settingpageaction/actions/SettingsActions.kt
@@ -1,0 +1,65 @@
+package com.github.nadiatarashkevich.settingpageaction.actions
+
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.options.ShowSettingsUtil
+
+class OpenAppearanceSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Appearance & Behavior")
+    }
+}
+
+class OpenKeymapSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Keymap")
+    }
+}
+
+class OpenEditorSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Editor")
+    }
+}
+
+class OpenPluginsSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Plugins")
+    }
+}
+
+class OpenVersionControlSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Version Control")
+    }
+}
+
+class OpenBuildSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Build, Execution, Deployment")
+    }
+}
+
+class OpenLanguagesSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Languages & Frameworks")
+    }
+}
+
+class OpenToolsSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Tools")
+    }
+}
+
+class OpenExternalToolsSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "External Tools")
+    }
+}
+
+class OpenLiveTemplatesSettingsAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        ShowSettingsUtil.getInstance().showSettingsDialog(e.project, "Live Templates")
+    }
+}